claude-orchestrator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-orchestrator"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk",
    "pyyaml",
    "ruff>=0.15.10",
    "watchdog",
    "requests",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.10",
]

[tool.hatch.build.targets.wheel]
packages = ["agents", "claude_orchestrator"]
